requirements.txt
streamlit==1.41.0
openai>=1.50.0
supabase>=2.0.0
rank-bm25>=0.2.2
numpy>=1.26.0
python-dotenv>=1.0.0